fix: fix reference list item icon alignment
* Fix broken HMR.
* Change label item buttons to icon buttons.
