+++ package.json
@@ -46,13 +46,14 @@
     "@testing-library/react": "^16.0.1",
     "@testing-library/user-event": "^14.5.2",
     "@types/d3": "^7.4.3",
-    "@types/react": "^18.3.5",
-    "@types/react-dom": "^18.3.0",
+    "@types/react": "^19.2.7",
+    "@types/react-dom": "^19.0.0",
     "@types/styled-components": "^5.1.34",
     "@types/superagent": "^8.1.9",
     "@vitejs/plugin-react": "^5.0.4",
     "@vitest/eslint-plugin": "^1.3.20",
     "babel-plugin-react-compiler": "^1.0.0",
+    "baseline-browser-mapping": "^2.9.11",
     "eslint": "^9.39.2",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-react": "^7.37.5",
