+++ src/sequences/components/SequenceButtons.tsx
@@ -2,7 +2,7 @@ import { useUrlSearchParam } from "@app/hooks";
 import CloseButton from "@base/CloseButton";
 import IconButton from "@base/IconButton";
 import { useGetActiveIsolateId } from "@otus/hooks";
-import { useCurrentOtuContext } from "@otus/queries";
+import { useCurrentOtuContext } from "@otus/CurrentOtuContext";
 import { DownloadLink } from "@references/components/Detail/DownloadLink";
 import { ReferenceRight, useCheckReferenceRight } from "@references/hooks";
 import styled from "styled-components";
